requirements.txt
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
numpy>=1
setuptools>=68
openreview-py>=1.0.0
openai>=1.0.0